完善自定义表单

+++ src/app/pages/sys/components/role/role.component.ts
@@ -68,13 +68,17 @@ export class RoleComponent implements OnInit, AfterViewInit {
   // 删除选择的角色
   onDeleteRole() {
     const that = this;
-    that._state.notifyDataChanged('delete.confirm', `${that.selectedRole.role_name}角色`);
-    that._state.subscribe('delete.confirm.result', (result) => {
-      that.roleService.delete(that.selectedRole.role_id).then(() => {
-        _.remove(that.roles, r => r['role_id'] === that.selectedRole.role_id);
-        that.selectedRole = null;
-      });
-    });
+    const confirm = {
+      message: `${that.selectedRole.role_name}角色`,
+      callback: function () {
+        that.roleService.delete(that.selectedRole.role_id).then(() => {
+          _.remove(that.roles, r => r['role_id'] === that.selectedRole.role_id);
+          that.selectedRole = null;
+        });
+      },
+    };
+
+    that._state.notifyDataChanged('delete.confirm', confirm);
   }
 
   onSelectedRole(role) {
